// denisbykov/progress-bar/types/tprogress-bar.ts
type TProgressBar = {
  fillColor?: string;
  backgroundColor?: string;
  children?: any;
  borderColor?: string;
  borderWidth?: string;
  borderStyle?: string;
  borderRadius?: string;
  height?: string;
  progress?: number;
  isAnimate?: boolean;
  fontSize?: string;
  direction?: "ltr" | "rtl";
};

export default TProgressBar;